+++ src/Spells/OdinFury.h
@@ -7,7 +7,7 @@
 
 class OdinFury: public Hechizo{
 public:
-	OdinFury(int costPM, float tCast, float tCoolDown, float optHP, float optMP);
+	OdinFury(float costPM, float tCast, float tCoolDown, float optHP, float optMP);
 	void Lanzar(Player* p);
 private:
